Moved 'jobsForm' inside formScreen, this way is easier to controll the key and validate the jobs form.
